board backend 리팩토링 및 프론트와 연결 수정

See merge request s09-webmobile1-sub2/S09P12A505!90

+++ server/Strange505/src/main/java/com/example/Strange505/board/service/CommentServiceImpl.java
@@ -4,15 +4,15 @@ import com.example.Strange505.board.domain.Article;
 import com.example.Strange505.board.domain.Comment;
 import com.example.Strange505.board.dto.CommentRequestDto;
 import com.example.Strange505.board.dto.CommentResponseDto;
+import com.example.Strange505.board.exception.NotAuthorException;
 import com.example.Strange505.board.repository.ArticleRepository;
 import com.example.Strange505.board.repository.CommentRepository;
 import com.example.Strange505.user.domain.User;
 import com.example.Strange505.user.repository.UserRepository;
-import com.example.Strange505.user.service.AuthService;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -20,7 +20,7 @@ import java.util.List;
 
 @Slf4j
 @Service
-@Transactional
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class CommentServiceImpl implements CommentService {
 
@@ -29,10 +29,11 @@ public class CommentServiceImpl implements CommentService {
     private final UserRepository userRepository;
 
     @Override
-    public void createComment(Long userId, CommentRequestDto dto) {
+    @Transactional
+    public void createComment(CommentRequestDto dto, String email) {
         Article article = articleRepository.getReferenceById(dto.getArticleId());
         Comment parent = commentRepository.findById(dto.getParentId()).orElse(null);
-        User user = userRepository.findById(userId).orElseThrow(()->new RuntimeException("댓글 생성 중: 사용자 아이디 없음"));
+        User user = userRepository.findByEmail(email).orElseThrow();
         Comment comment = null;
         if (parent == null) {
             comment = Comment.builder()
@@ -76,7 +77,9 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public List<CommentResponseDto> getCommentByUser(Long userId) {
+    public List<CommentResponseDto> getCommentByUser(String email) {
+        User user = userRepository.findByEmail(email).orElseThrow();
+        Long userId = user.getId();
         List<Comment> list = commentRepository.searchByUser(userId);
         List<CommentResponseDto> dtoList = new ArrayList<>();
         list.stream().forEach(findByArticle -> dtoList.add(new CommentResponseDto(
@@ -87,21 +90,32 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public CommentResponseDto updateComment(Long id, CommentRequestDto dto) {
+    @Transactional
+    public CommentResponseDto updateComment(Long id, CommentRequestDto dto, String email) {
+        User user = userRepository.findByEmail(email).orElseThrow();
         Comment comment = commentRepository.findById(id).orElseThrow(()->new RuntimeException("Comment not found"));
-        comment.update(dto.getContent(), LocalDateTime.now());
-        Comment save = commentRepository.save(comment);
-        return new CommentResponseDto(save);
+        if (user.getId() == comment.getUser().getId()) {
+            comment.update(dto.getContent(), LocalDateTime.now());
+            Comment save = commentRepository.save(comment);
+            return new CommentResponseDto(save);
+        } else {
+            throw new NotAuthorException("작성자만 삭제 가능합니다.");
+        }
+
     }
 
     @Override
-    public void deleteComment(Long id) {
+    @Transactional
+    public void deleteComment(Long id, String email) {
+        User user = userRepository.findByEmail(email).orElseThrow();
         Comment comment = commentRepository.findById(id).orElseThrow(() -> new RuntimeException());
-        comment.remove();
-
-        List<Comment> removableCommentList = comment.findRemovableList();
-        log.info("removeList = {}", removableCommentList);
-        commentRepository.deleteAll(removableCommentList);
-        System.out.println("서비스단");
+        if (user.getId() == comment.getUser().getId()) {
+            comment.remove();
+            List<Comment> removableCommentList = comment.findRemovableList();
+            log.info("removeList = {}", removableCommentList);
+            commentRepository.deleteAll(removableCommentList);
+        } else {
+            throw new NotAuthorException("작성자만 삭제 가능합니다.");
+        }
     }
 }
